Convert Project and ProjectImageEntry to interfaces
Refactoring projects to make Project and ProjectImageEntry interfaces. There is a legacy implementation for backwards compatibility, and the beginning of a 'default' implementation for the future (although it does not yet work).
