+++ qupath-gui-fx/src/main/java/qupath/lib/gui/viewer/DragDropFileImportListener.java
@@ -48,6 +48,7 @@ import qupath.lib.images.ImageData;
 import qupath.lib.objects.hierarchy.PathObjectHierarchy;
 import qupath.lib.objects.hierarchy.TMAGrid;
 import qupath.lib.projects.Project;
+import qupath.lib.projects.ProjectFactory;
 import qupath.lib.projects.ProjectIO;
 import qupath.lib.gui.scripting.DefaultScriptEditor;
 
@@ -206,7 +207,7 @@ public class DragDropFileImportListener implements EventHandler<DragEvent> {
 				} else if (filesInDirectory.length == 0) {
 					// If we have an empty directory, offer to set it as a project
 					if (DisplayHelpers.showYesNoDialog("Create project", "Create project for empty directory?")) {
-						Project<BufferedImage> project = new Project<>(file, BufferedImage.class);
+						Project<BufferedImage> project = ProjectFactory.createProject(file, BufferedImage.class);
 						gui.setProject(project);
 						return;
 					}
